requirements.txt
django
djangorestframework
drf-spectacular
Pillow
gunicorn
django-storages
boto3
python-dotenv
google-cloud-storage
django-cloudinary-storage